RENAMING OF DNS ROLES
Renamed the dns_internal role to simply dns. In Docker testing I find
the need to reuse this role to simulate the external DNS service, which
I don't implement using Ansible in live as I simply use the one provided
by my ISP. However, the need to do this in Docker testing makes the use
of a role names specifically dns_internal incongruous.

Consequently, I have also renamed the dns_external role to dynamic_dns,
which is actually also more fitting to what it is about.