ADDED ROLE BASED BACULA FILESET CONFIGURATION FILES EN MASSE
Primary roles, not those that are only pulled in as dependencies, will
each have a bacula-fileset.conf file associated with them. These are
used by the backup service to build the required backup fileset for each
host based on the roles that are deployed to that host. This approach
has only passed basic testing at this stage.