REMOVE FUNCTIONALITY SPECIFIC TO THE CONTAINER ENVIRONMENT
When I'm testing these roles in a container environment I have
containers to simulate our ISP's DNS service and the SMTP service for a
user who is not one of our users nor one of the users for a customer we
provide the email service to. Neither of these are services that we
provide in live and so are not targets for our Ansible deployment in
live.

I had expanded roles to cater for these requirements. I've decided that
was the wrong approach. Instead anything specific to the container
environment only will be implemented directly in the playbooks for that
environment and not in this roles library.