MAKE SMTP ON OFFICE NETWORK PART OF THE HOST ROLES DEPLOYMENT
I had got this as one aspect of the home domain deployment. However we
use a common SMTP service on the office network for all domains, the
home domain and also the domains of our customers. So, it's more correct
actually to make this service part of the host roles deployment.