REMOVE REDUNDANT FILESET FILES All hosts are backup clients and the Bacula configuration folder is common across the Bacula client, Bacula director and Bacula storage daemon.
Database backups will be catered for via the export of databases from
the other roles that use one; for example, the WordPress role. Note that
the database role's FileSet file had been an empty placeholder. The
backup director role will need to change to cater for roles without a
FileSet file.