CREATE WORDPRESS_APACHE AND WORDPRESS_NGINX ROLES
The current wordpress role caters for wordpress deployment using Apache
or Nginx. I'm going to change to an arrangement whereby we have
wordpress_apache and wordpress_nginx roles, both of which have a
dependency on a wordpress role, which will be stripped back to the
common core. This will be mirrored for the domain_wordpress role.

In the first instance, this commit:

 - Moves wordpress/files/bacula-fileset.conf to both the new roles prior
   to changing it so that we can track changes.

 - Does the same for wordpress/templates/wordpress-backup.sh.j2 for the
   same reason. Note, in this case the file is copied to the files
   folder in the target roles and the .j2 suffix is dropped, since post
   editing it will no longer need to be a template.

 - Copies the wordpress/tasks/main.yml to both the new roles, again
   prior to changing it in all three locations so that we can track
   changes.