MOVE CREATE TASKS TO MAIN TASKS WHERE LOOPING WITHIN ROLE REMOVED
The site_wordpress and site_wordpress_database roles had a main task
list that just provided a loop import of a create task list. I'm going
to provide the loop in playbooks, outside of the role. So, that create
task list can now become the main task list. In the first instance I'm
just renaming the files so that changes after that point can then be
tracked.