REPLACE META DEPENDENCIES WITH IMPORTS IN MAIN TASK LISTS
I want to have other task lists within role, for example delete task
lists, which do not invoke a role's dependenices. So, remove the meta
depdencies and replace them with import role tasks in the main task
list. That way they're not invoked when the other task lists are used.