USE PLUGIN NAMES AS KEYS IN THE LIST OF PLUGINS FOR A WORDPRESS SITE
Previously I configured plugins as an array of dictionary objects with
the plugin "name" being one of the attributes of each dictionary object.
The code is much simplified if they are instead configured as a
dictionary object with each key within that object being a plugin name.

Include a couple of unrelated and minor tweaks, one to wrap a line and
another because an include role rather than an import role was needed.