CHANGES WHEN USING DYNAMIC DNS ROLE TO DEPLOY FOBV DOMAIN CONFIG
This was the first time we had deployed the dynamic DNS configuration
for a domain.

I started by creating a separate domain_dynamic_dns role but then
changed to using a domain.yml task list within the existing role. I
think that I will adopt this approach going forwards for other roles,
which would lead to dropping the domain and site roles.

Also added explanatory comments and changed the deployment of the
dynamic DNS script so that it's called by perl as a source file rather
than being defined as executable.