RENAME TASK LIST DOMAIN.YML TO PROJECT.YML IN DYNAMIC_DNS ROLE
The way that this role is currently structured it expects there to be a
one-to-one correlation between projects and domains. However, this isn't
always the case. For varilink.co.uk we have a project for the website
and a separate project for our back-office services. So, I'm going to
change this role to reflect that.

First of all, renaming tasks/domain.yml to tasks/project.yml, as that's
more appropriate in light of this and comitting so that subsequent
changes to tasks/project.yml can then be tracked.