RENAMING OF SITE TASK LIST FILES
Only use underscore at the start of the folder to ensure it appears at
the top of file listings. After that use hypens as I normally do. This
remains an invalid Ansible role name.

Shorten the names of the task list files within the folder. If they're
in that folder then we already know that they pertain to WordPress
sites.