EXTENSIVE REWORKING OF WORDPRESS ROLE
The wordpress role has been extensively reworked and restested in
conjunction with the wordpress_nginx role and tested using the
fobv-ansible repository. The FoBV website is deployed as an Nginx based
WordPress site.

I now want to continue testing the revised wordpress role but now using
the website-ansible repository, which manages the deployment of the
Varilink Computing Ltd website. That website is deployed as an Apache
based WordPress site, so now I plan to change and test the wordpress
role further but this time in conjunction with the wordpress_apache
role.