CHANGES AFTER BASIC CREATE/DELETE SITE TESTING WITH VARILINK WEBSITE
The testing was to run through a cycle of creating and deleting the
Varilink website in the dev environment using the website-ansible
repository. As such the affected roles are reverse_proxy and
wordpress_apache, since this is an Apache based WordPress site. There
are also some updates to the related site task lists, including
splitting off the tasks to select a single subdomain on a host for
multiple subdomains of the same domain to make these tasks reusable
across projects.