CHANGES AFTER BASIC CREATE/DELETE SITE TESTING WITH FOBV WEBSITE
The testing was to run through a cycle of creating and deleting The
FoBV website in the dev environment using the fobv-ansible repository.
As such the affected role is wordpress_nginx, since this is an Nginx
based WordPress site. There are also some updates to the related site
task lists.