IMPROVEMENTS TO THE DYNAMIC_DNS ROLE
Made the update to crontab entries much more precise. Now it will only
enact the changes that it has to make if the stride is changed and won't
report unless it actually has changed something.

Enable the ability for the dynamic-dns.pl script to run without the
dns-linode-credentials.ini file in place, in which case it won't attempt
any API calls but will report the record file that it "processes" (it
doesn't actually do anything with them in this scenario). This is to
enable a level of testing without having to have the Linode
configuration in place.