CHANGES ARISING FROM DEPLOY/DELETE CYCLE TESTING OF SSL ENABLED SITE
Changes made when undertaking testing by cycling through the deploy and
delete of a WordPress site that uses a Let's Encrypt issued SSL
certificate. Key is to get the website up and running without SSL so
that the certificate requests can then be authenticated using the
certbot "webroot" method. The certbot integration is automated for the
first time here.

I've also introduced the idea of two-tier SSL indicator variables,
host_uses_ssl (the host does or does not support SSL) and
wordpress_site_uses_sll (a WordPress site does or does not use SSL).
This approach will need to be applied to the wordpress_nginx role as the
testing here was with the wordpress_apache role.

The change includes some simplication in the creation of the WordPress
config file and the changes needed in that file for WordPress running
under Apache to be behind a reverse proxy with SSL enabled.