FIX AND IMPROVEMENT TO STEPS THAT CREATE A NEW OPERATING SYSTEM USER
We were supplying the plain text password in the ansible.builtin.user
which is wrong, that module requires that you give it the password hash.

Furthermore, since the md5 hashing algorithm uses a random salt if you
don't provide one, we were getting a change reported every time, even if
the password hadn't changed just because the salt had changed. So now
we check the password against the stored hash using the same salt that's
in the stored hash to determine where the password has actually changed
or not. Then we only use the module if either the user is new or the
password has changed for an existing user.