CORRECTLY USE SHELL RC TO DETERMINE SUCCESS OR ERROR
I was suppressing errors on the task to determine if WordPress is
installed or not for a WordPress site. That's because, just like grep,
if it doesn't find that WordPress is installed the return code to the
shell is one. Much better to add a failed_when that considers both
return code 0 and return code 1 as not a fail.