CATER FOR BACKUP STORAGE DAEMON THAT IS LOCAL TO THE BACKUP DIRECTOR
If the variable backup_storage_host isn't defined then we assume that's
because it's running local to the backup director, so the backup
director doesn't need to be told where the backup storage host is. That
being the case the backup storage host can be accessed on the localhost
network interface.