PROVIDE DEFAULT VALUE FOR MAIL_USES_CA IN MAIL_EXTERNAL ROLE
This role now needs to be sure that this variable is defined as it is
generating self-signed certificates itself rather than receiving them
from the mail_certificates role when mail is NOT using a certification
authority. I had a failure in testing because this variable hadn't been
set in either the inventory or playbook, so provide a default value
here.