REMOVE SELF-SIGNED CERTIFICATES OPTOIN FROM MAIL_CERTIFICATES ROLE
If a domain is going to use a self-signed certificate then now we
generate it in-situ on the host for the mail_external role. We only need
to generate them on the host of this role if we're using a certification
authority as it is the host of this role that can use the Let's Encrypt
DNS validation method.