FLEXIBILITY IN DEFINING HOSTS TO BACKUP
We were defining the hosts to backup via one option only, the definition
of a backup_clients group. Now we've added the means to backup the
default all group or everything in the all group excepts hosts in a
backup_exceptions group. You can choose any one of those three means to
define the hosts to backup.