MOVE CREATE/DELETE SITE TASKS IN DATABASE, DNS AND REVERSE PROXY ROLE
Consistent with the previous commit these task lists are moved from
within the site/ subfolder from where they were imported by the within
role create-sites and delete-sites task lists to be in the route tasks/
folder as create-site and delete-site task lists. They will now need to
be imported by the wordpress and wordpress_apache roles and so need to
sit at the top-level within the tasks folder of their own roles.