WORDPRESS CONFIG FOR BEING BEHIND A PROXY IN THE WORDPRESS APACHE ROLE
The wordpress_apache role is implemented behind a reverse proxy. This
requires some settings in the WP config file. If the WordPress site uses
SSL then we must set $_SERVER['HTTPS']='on' as WordPress won't know that
HTTPS is implemented on the proxy. We also need to set $_['REMOTE_ADDR']
to $_SERVER['HTTP_X_REAL_IP'] so that, for example, the remote IP
address is shown on comments and not the IP address of the reverse
proxy.

While we're at it I also changed the indentation of tasks within a block
to my general standard.