ADD DELEGATIONS FOR WORDPRESS ROLE COORDINATION OF SITE CREATION
I have moved to an approach here whereby only the wordpress roles do the
site loop. They then import the database, dns and reverse_proxy create
site tasks as needed.

Obviously, those database, dns and reverse_proxy tasks may need to be
executed on different hosts to the wordpress host itself, so we use
delegation if that's the case. The need to delegate is indicated by the
values of the wordpress_site_database_host, wordpress_site_dns_host and
wordpress_site_reverse_proxy host variables.

For the database and reverse_proxy tasks, if those host variables are
set then the tasks are delegated to the indicated host. If they're not
set then they're delegated to inventory_hostname, effectively "delegate
to self". The dns tasks are only run if wordpress_site_dns_host is
defined, because we want to skip that for hosts on the external network
in the live environment.