ADD IMPORTS FOR DATBASE, DNS AND REVERSE PROXY TASKS IN WORDPRESS ROLES
Now that the wordpress roles are driving the creation of sites they need
to import the create site tasks from the database, dns and reverse proxy
roles.

Note that the reverse proxy role import is in the wordpress_apache role,
since it's only applicable to our stack then. It's not applicable when
we use the wordpress_nginx role. The database and dns imports are
through the shared wordpress role, since they're applicable to both of
those architectures.