EXPAND NAMES OF SERVICE RELATED TASK LISTS
I've expanded configure and install to configure-service and
install-service, which is better I think because it clearly
differentiates to site, domain, project, etc. level task lists. Right
now just done this for the dns, monitor_client and monitor_server roles
as I was working on them but I will extend this practice to all roles in
time.