ADD SOME EXPLICIT SERVICE MANAGEMENT STEPS IN TO RECENTLY ADDED ROLES
It seems that in the recently added monitoring roles the services are
not automatically started. Also I made the change recently to move
handlers to service.yml rather than main.yml files so that they wouldn't
apply in my "Services - Docker" test environment, which is based on
Docker containers that don't have a service manager within them.

These explicit references to the service manager are an interim measure
used when applying and testing the new backup web and monitoring
services in my live server estate. I'll have to change this approach in
the future to be flexible across both this server environment and also
my "Services - Docker" container based testing environment.