RENAME OR DELETE SERVICE HANDLER FILES AND ADD NONE HANDLER FILES
Rename multiple service.yml handler files to main.yml so that they will
automatically be in the scope of role imports or includes unless
specified otherwise. Add none.yml handler files to give the option of
importing or including roles without bringing the handlers in the
main.yml file. Delete the handler for the reverse_proxy role as it no
longer needs its own handlers now that it imports those of the nginx
role.