CLEAR OUT OF REVERSAL PROXY ROLE NOW THAT IT IMPORTS FROM NGINX ROLE
Several of the files in the reverse_proxy role are redundant now that
the reverse_proxy imports from the new nginx role in it's main.yml task
list, which is also modified as a result. The equivalent files in the
nginx role are direct replacements for those in the reverse_proxy role
that are deleted in this commit.