CORRECT ISSUE ARISING IN THE BACKUP DIRECTOR CREATE CATALOG TASKS
Due to the introduction of the backup_web role we now need to define the
backup_database_host variable even if the backup database is cohosted
with the backup director. Thus the check in the backup director role to
determine if the local socket can be used to access the database needs
to not only check that backup_database_host is defined or not but also,
if it is defined, whether it matches the current hostname.