DON'T CREATE A DNS RECORD FOR EXTERNALLY HOSTED WEBSITES
This is a quick workaround to avoid an error when using this roles
library to deploy a project's website. It will have to be improved upon
as when I test this roles library with the Varilink Services - Docker
repository I need to create DNS records for containers that are
simulating hosts on the external network. Also, I'd like to expand this
role at some point so that it caters not also for my internal use of
Dnsmasq but also configures DNS entries with our VM hosting partner when
needed.