GIVE TASK FILE TO IMPORT NGINX HANDLERS INTO REVERSE PROXY ROLE NEW NAME
The previous name was "create-site-handlers.yml". It's now been named
"import-nginx-handlers.yml", which is more descriptive of what it
actually does. It is also now imported by the "delete-site.yml" task
list so a name specific to it being imported by the "create-site.yml"
task list no longer makes sense.