ADD PHP XML PACKAGE IN WORDPRESS ROLE AND NOTIFY APACHE2 CONFIG CHANGE
I have noticed that the php-xml package is required by at least one of
my current websites so I might as well include it in the WordPress
service.

It seems that whenever one of these PHP packages is installed then the
Apache2 configuration must be reloaded for mod_php to pick the new
package up, hence I've put the associated notify on the package install.
This may break in existing playbooks where the relevant handlers are not
loaded at that point but I'll resolve that issue if it exists on a
case-by-case basis.