ADD OPTION TO UPLOAD MEDIA TO WORDPRESS SITES FROM THE CONTROLLER
Before this, if we wanted to upload media to a WordPress site in WP-CLI
scipts then that media had to be included in a project theme file.
However not all projects have their own theme and in any case it's
redundant to have the images in the theme after they have been uploaded
to the media library. So, now we can take them from the project's
WordPress repository on the controller without having to add them to a
project's theme.