ENABLE WP-CLI TO HARD FLUSH REWRITE RULES
WP-CLI scripts that are update the permalink structure have been
affected by a bug that meant that the new permalink structure that they
set was not taking effect. After some invesigation, I realised that was
because they did not have the `--hard` flag set when they ran the WP-CLI
`rewrite flush` command. For this flag to take effect however, we also
need to add a `config.yml` file in the `.wp-cli` directory within the
`www-data` user's home directory with some required configuration
settings in it. This change makes sure that this is created when the
base WordPress hosting service is deployed.