REMOVE ORCHESTRATION OF ROLES FROM CREATE/DELETE WORDPRESS APACHE SITE
I have concluded that it going to be cleaner and clearer what's going on
if role task list do NOT in general orchestrate by importing and/or
including other task lists, from the same role or other roles. Instead
I am going to do this within the common playbooks and/or common task
lists that are imported and/or included into those playbooks, all within
the Varilink Ansible Playbooks repository.