CHANGE DEFAULT RSYNC PATH TO MATCH THAT USED WHEN NONE IS SPECIFIED
Through examining debug output, I've established that if you use the
Ansible synchronize module and you do not specify a rsync_path then the
rsync path defaults to `sudo -u www-data sync` (within the context of
the `create-site.yml` task list in the WordPress role. So I've changed
the Jinja templating that sets rsync_path to make that the default if
no rsync_path is provided.