FALLBACK "USE IPIFY" OPTION FOR THE DYNAMIC DNS ROLE
The dynamic DNS services uses a feature of the Linode API whereby if you
use the string '[remote_addr]' in the JSON that you post it will refer
to your current, externally facing IP address as Linode sees it.
Recently, we had an incident whereby this was returning an internal (to
Linode) 192 address instead, perhaps the internal address of a firewall
or router interface.

Linode fixed the problem PDQ but in the meantime I implemented a
workaround whereby the dynamic DNS role can be temporarily instructed to
use the ipify API instead to get the current, external IP address. We
should retain that option in case there are ever any similar problems
again in future.