CATER FOR DEBIAN BOOKWORK RELEASE IN THE INFLUXDB ROLE
We added a server to our host estate using the latest stable, Debian
bookworm release. However, InfluxData have yet to add a packages
specifically for the bookworm distro to their package repository. As a
workaround the influxdata role is tweaked to use the InfluxData packages
for Debian bullseye if the target host is based on Debian bookworm.