ADD REPEATABLE AUTOMATION TO THE INSTALLATION OF THE MONITOR DASHBOARD
Previously I just installed a base service and then configured in
manually. Of course the current configuration is backed up and so could
be restored from backup, but I want to capture the configuration as part
of the installation process. The work here is still draft and has been
used from the services-ansible repository to test some basics of this
repeatable automation. I'm committing this work in progress so that it
can be pulled into the services-docker repository to be worked on and
tested further there.