ADDITION OF DISABLE AND STOP TASK LISTS TO ROLE
These have been used in combination with a new playbook for stopping
services (and optionally disabling them as well) during the migration of
services from the old hub server to its replacement. The were used for
stopping and disabling services on the old server and the new server
took over. We will need to add disable and stop task lists to other
roles if we stick with and extend this idea.