ALLOW FOR HOSTS THAT AREN'T ALIASED WHEN CONFIGURING THE BACULA DIRECTOR
This change arose while migrating the hub role from one host to another
host. When doing this, I realised that it's useful to retain the hub
role as an alias for the old host while referring to the new host
directly. This was so that I could select old or new host for playbook
actions for a period.