CATER FOR BOOKWORM IN CALENDAR ROLE AND ENABLE THE SERVICE ON INSTALL
Some testing has exposed that when the role is installed the service is
not enabled and started without further configuration being required to
do this. So, I've added that configuration. Furthermore it turns out
that this configuration differs between Radicale version two and three.
Version two is installed with buster and version three is installed with
bullseye and bookworm.