CONFIGURE RESOLV.CONF FILE PROPERLY FOR DNS SERVERS
A host acting as a DNS server should resolve DNS queries via itself in
combination with the upstream servers configured in Dnsmasq. So, we
configure /etc/resolv.conf appropriately in the dns role rather than
configuring it via the dns_client role for these hosts. This requires
that the host selection patterns be set accordingly so that servers that
are providing a DNS service aren't configured via the dns_client role.