REMOVE ROLE PATH FROM NAME WHEN IMPORTING ROLES WITHIN THIS LIBRARY
When I test this role library using the "Services - Docker" repository I
use a library of "wrapper" roles to customise their behaviour for the
container environment that "Services - Docker" uses. At some point this
prompted me to use the Ansible role_path variables when importing one
role in this library from another role in this library.

Actually that's not necessary, since it's only this role library in the
role path when using "Services - Docker". The roles in the wrapper role
library are referenced by absolute path within the containers. So, we
can remove these references to role_path and just use the role name
alone.