CHANGE TO VOLUME RETENTTION FOR THE CATALOG COPY POOL
I was sometimes getting two catelog copy jobs running on the same night
because a copy volume had expired just prior to the original volume
expiring and getting reused. This meant that Bacula copied the catalog
volume created that night plus and old one, for which it now knew no
valid copy was in place. Added a day to the retention period for the
copy volume to prevent this scenario from happening.